--- README.md
@@ -1,6 +1,0 @@
-#Wub
-
-Минималистичный фрэймворк для построения неинтерактивных приложений, обработчиков запросов. Для обучения веб-разработке, для прототипов. И для провокаций в команде высококлассных специалистов :-).
-
-Основная идея заключается в том, что код фрэймворка в вашем приложении принадлежит вам. Никаких тайных знаний. Объём кода такой, что вы можете разобраться в нём за один подход. А когда придёт время оправдать вашу высокую зарплату объёмом кода в репозитории, перейдя на симфони, выбросить эту поделку будет легко как никогда!
-
