+++ bonk/HttpResponse.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace frm;
+namespace bonk;
 
 class HttpResponse implements Response {
 
@@ -10,9 +10,12 @@ class HttpResponse implements Response {
 		private string $body,
 	) {}
 
-	public function send($file): int {
+    /**
+     * @inheritDoc
+     */
+	public function send($file) {
 		if (\headers_sent()) {
-			throw new Err('headers already sent');
+			throw Exception::system('headers already sent');
 		}
 
 		\http_response_code($this->code);
@@ -25,10 +28,18 @@ class HttpResponse implements Response {
 			\header($header, true);
 		}
 		
-		\fwrite($file, $this->body);
+		$r = \fwrite($file, $this->body);
+        if ($r === false) {
+            throw Exception::system("can't send response");
+        }
 
-		return 0;
+        return 0;
 	}
+
+    public function setOutput(string $output): self
+    {
+        return $this->setBody($output);
+    }
 
 	public static function notFound(string $message, $context = []): self {
 		return new self(404, [], $message);
