+++ docs/arch.md
@@ -14,26 +14,28 @@ actor User
 User -> index.php
 index.php -> App : run(rq)
 App -> Router : route(rq.url)
+Router --> App: handlerId
 App -> Registry : get(handlerId)
 
 database "./registry/*" as fs
 
 Registry -> fs : require
-fs -> Registry : valueBuilder
-Registry -> ValueBuilder : build()
-Registry -> App : handler
+fs --> Registry : factory
+Registry -> Factory : __invoke(registry)
+Factory --> Registry: handler
+Registry --> App : handler
 App -> handler : __invoke(rq, rs)
-handler -> App : rs
+handler --> App : rs
 App -> Response : send()
-App -> index.php
-index.php -> User
+App --> index.php: exitCode
+index.php --> User
 
 @enduml
 ```
 
 ## Внедрение зависимостей
 
-Фреймворк содержит решение основанное на реестре Registry и построителе значений ValueBuilder. Реестр ищет скрипты-фабрики значений в файловой системе (return из подключённого файла), если в результате получает ValueBuilder, то вызывает его метод build, который получает зависимости, вызывая Registry::get (косвенная рекурсия), и строит значение.
+Фреймворк содержит решение основанное на реестре Registry и фабрике зависимостей Factory. Реестр получает зависимости, подключая файл скрипта-фабрики (скрипт возвращает фабрику через return).
 
-Любой сервис или контроллер описывается как ValueBuilder и может иметь зависимости или быть чей-то зависимостью.
+Любой сервис или обработчик запроса является экземпляром Factory и может иметь зависимости или быть чьей-то зависимостью. Например, объект приложения App находит нужный обработчик запроса через реестр.
