+++ readme.md
@@ -1,0 +1,62 @@
+# Bonk! -- ультимативный фрэймворк
+
+Это очень простой фрэймворк для построения неинтерактивных приложений,
+обработчиков запросов. Отлично подойдёт для обучения веб-разработке, для прототипов.
+А так же для провокаций в команде высококлассных специалистов :-)
+
+Основная идея заключается в том, что код фрэймворка в вашем приложении принадлежит вам.
+Никаких тайных знаний и тёмных углов. Объём кода такой, что вы можете
+разобраться в нём за один подход.
+
+
+## Использование
+
+### Создание каркаса приложения
+
+1. Напишите точку входа (например index.php), которая инстанцирует приложение (`bonk\App`)
+и запустите его с запросом, полученным из окружения.
+
+```php
+<?php declare(strict_types=1);
+
+namespace bonk;
+
+require __DIR__ . '/vendor/autoload.php';
+
+$router = new Router([]);
+$registry = new Registry([]);
+$app = new App($router, $registry);
+$rq = HttpRequest::fromGlobals($_SERVER, $_GET, $_POST, $_FILES, $_COOKIE, 'php://input');
+$rs = HttpResponse::ok();
+$exitCode = $app->run($rq, $rs, fopen('php://output', 'w'), fopen('php://stderr', 'w'));
+exit($exitCode);
+```
+
+2. Создайте папку для фабрик обработчиков запроса и добавьте её в реестр (`bonk\Registry`).
+
+3. Создайте папку для фабрик зависимостей и тоже добавьте её в реестр.
+
+```php
+$registry = new Registry([
+	'web/' => __DIR__ . '/web',
+	'' => __DIR__ . '/factories',
+]);
+```
+
+4. Сконфигурируйте роутинг.
+
+```php
+$router = new Router([
+	'~GET /~' => 'web/index.php',
+    '~GET /hello~' => 'web/helloWorld',
+]);
+```
+
+[Код целиком](./tests/example/app/web.php) можно подглядеть в примере приложения.
+
+### Написание кода вашего приложения
+
+Отделите специфичный для вашего приложения код от кода каркаса. Фабрики обработчиков
+запросов и фабрики зависимостей лучше писать в нэймспейсе `bonk`. Но остальной ваш
+код не должен ничего знать про фрэймворк.
+
