--- src/Err.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace frm;
-
-class Err extends \Exception {
-	const CODE_SYSTEM = 1;
-}
