+++ tests/example/app/cli.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace frm;
+namespace bonk;
 
 require __DIR__ . '/../../../vendor/autoload.php';
