+++ tests/example/app/factories/articles/service.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
-namespace frm;
+namespace bonk;
 
-return make(function(\example\articles\Repo $repo) {
+return _(function(\example\articles\Repo $repo) {
 	return new \example\articles\Service($repo);
 })->with(repo: 'articles/repo.php');
