--- tests/example/app/handlers/helloWorld.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace frm;
-
-return fn() => function() {
-	return 'Hello world!';
-};
