+++ tests/example/app/web/helloWorld.php
@@ -1,0 +1,5 @@
+<?php declare(strict_types=1);
+
+namespace bonk;
+
+return fn() => 'Hello world!';
