+++ tests/example/public/index.php
@@ -1,5 +1,5 @@
 <?php declare(strict_types=1);
 
-// php -S localhost:8080
+// php -S localhost:8080 index.php
 
 require __DIR__ . '/../app/web.php';
